Move court integration test to `integration.rs`
